test(alloy): inline nonce regression tests in module
